Mainstream Framework Integration/Flask/requirements.txt
Flask
aspose-cells-python
